Added Room Database. (#5)
* Added Room Database.

* Fixed repository pattern.

+++ app/src/main/java/ru/knowledge/mtstetaproject/MainActivity.kt
@@ -27,7 +27,7 @@ class MainActivity : AppCompatActivity(), StartFragmentDetailsListener {
         bottomNavigationView.setupWithNavController(navController)
     }
 
-    override fun onStartFragmentDetails(movieId: Int) {
+    override fun onStartFragmentDetails(movieId: Long) {
         val args = bundleOf(MovieDetailsFragment.MOVIE_ID to movieId)
         navController.navigate(R.id.action_movies_to_details, args)
     }
